refactor(frontend): add ItemListPage with routes
